Kept working backend... 	ε(´סּ︵סּ`)з
The question displayed is now being taken from the actual game you're playing, no longer hard-coded. The emoji displayed when you're in the quiz is also accurate. Next step is to keep working in the "endRound"-function to send it further to keep rendering the quiz.
Also added points counter to back-end, and also a question-counter.