Fixed bugs with ID and localStorage
localStorage wasn't clearing properly when leaving a game, may result in bugs, so localStorage is now cleared of everything except for players ID when renderHome function executes.

Also used endGame event for displaying finalScore screen accidentally, so endGame is now used for displaying end-view, and quitGame is now used for terminating a game when host leaves.